requirements.txt
playwright>=1.48.0
PyYAML>=6.0.2
pytest>=8.3.0
